Create cadfPairTrade
I use this code on Spyder within Anaconda frame. It returns promising pairs to be studied further. We then use WorkOnPair code to study further.
Some assets in markets are related to each other. We make use of this relation to profit from their relation. Assume two stocks A and B, a pair, and they are in the same industry (VISA and MASTERCARD). Their performance vary along time and the spread between them goes wide and then meet back at zero spread. If this happens then there is stationarity between them. We balance our positional hedge ratio between them then short the overperforming one and long the underperforming one wait for the spread to get narrower so we make profit. Our expectation is that the overperformer will lose stamina and underperform from now on, or vice versa for the underperformer, or both cases.